[#IOPLT-1056] New Auth & Identity product, auth API Group, routing (#1563)
Add a new API Product to the Platform proxy for domain `auth and
identity`.
Define a new API Group with the API definined into `io-backend` for the
`auth and identity` domain.
Add routing for the API into the Application Gateway.
Add RBAC for Auth & Identity team admin group and CD identity to the
proxy.